+++ packages/zosfiles/src/methods/copy/Copy.ts
@@ -58,6 +58,7 @@ export class Copy {
     ): Promise<IZosFilesResponse> {
         ImperativeExpect.toBeDefinedAndNonBlank(options["from-dataset"].dsn, "fromDataSetName");
         ImperativeExpect.toBeDefinedAndNonBlank(toDataSetName, "toDataSetName");
+        const safeReplace: boolean = options.safeReplace;
 
         if(options["from-dataset"].dsn === toDataSetName && toMemberName === options["from-dataset"].member) {
             return {
@@ -71,6 +72,15 @@ export class Copy {
         const newDataSet = !targetDataSetExists;
         if (newDataSet) {
             await Create.dataSetLike(session, toDataSetName, options["from-dataset"].dsn);
+        }
+        else if(safeReplace) {
+            if (options.promptFn != null) {
+                const userResponse = await options.promptFn(toDataSetName);
+
+                if(!userResponse) {
+                    throw new ImperativeError({ msg: ZosFilesMessages.datasetCopiedAborted.message });
+                }
+            }
         }
         if(!toMemberName && !options["from-dataset"].member) {
             const sourceIsPds = await this.isPDS(session, options["from-dataset"].dsn);
@@ -265,7 +275,7 @@ export class Copy {
                 /*
                 * If the source is a PDS and no member was specified then abort the copy.
                 */
-                if((sourceDataSetObj.dsorg == "PO" || sourceDataSetObj.dsorg == "POE") && sourceMember == undefined){
+                if(sourceDataSetObj.dsorg.startsWith("PO") && sourceMember == undefined){
                     throw new ImperativeError({ msg: ZosFilesMessages.datasetCopiedAbortedNoPDS.message });
                 }
             }
@@ -309,7 +319,7 @@ export class Copy {
                     targetDataSetObj = TargetDsList.apiResponse.items[dsnameIndex];
                     targetFound = true;
 
-                    if((targetDataSetObj.dsorg == "PO" || targetDataSetObj.dsorg == "POE") && targetMember == undefined)
+                    if(targetDataSetObj.dsorg.startsWith("PO") && targetMember == undefined)
                     {
                         throw new ImperativeError({ msg: ZosFilesMessages.datasetCopiedAbortedTargetNotPDSMember.message });
                     }
@@ -339,11 +349,11 @@ export class Copy {
                 * If this is a PDS but the target is the sequential dataset and does not exist,
                 * create a new sequential dataset with the same parameters as the original PDS.
                 */
-                if((createOptions.dsorg == "PO" || createOptions.dsorg == "POE") && targetMember == undefined){
+                if(createOptions.dsorg.startsWith("PO") && targetMember == undefined){
                     createOptions.dsorg ="PS";
                     createOptions.dirblk = 0;
                 }
-                else if(targetMember != undefined &&  (createOptions.dsorg != "PO" && createOptions.dsorg != "POE"))
+                else if(targetMember != undefined &&  !createOptions.dsorg.startsWith("PO"))
                 {
                     createOptions.dsorg ="PO";
                     createOptions.dirblk = 1;
@@ -412,21 +422,27 @@ export class Copy {
             storclass: targetOptions.targetStorageClass,
             mgntclass: targetOptions.targetManagementClass,
             dataclass: targetOptions.targetDataClass,
-            dirblk: parseInt(dsInfo.dsorg == "PO" || dsInfo.dsorg == "POE"  ? "10" : "0")
+            dirblk: parseInt(dsInfo.dsorg.startsWith("PO") ? "10" : "0")
         }));
     }
 
     /**
-     *  Private function to convert the ALC value from the format returned by the Get() call in to the format used by the Create() call
+     * Converts the ALC value from the format returned by the Get() call to the format used by the Create() call.
+     * @param {string} getValue - The ALC value from the Get() call.
+     * @returns {string} - The ALC value in the format used by the Create() call.
      */
-    private static convertAlcTozOSMF( zosmfValue: string): string {
+    private static convertAlcTozOSMF(getValue: string): string {
         /**
          *  Create dataset only accepts tracks or cylinders as allocation units.
          *  When the get() call retreives the dataset info, it will convert size
          *  allocations of the other unit types in to tracks. So we will always
          *  allocate the new target in tracks.
         */
-        return "TRK";
+        const alcMap: Record<string, string> = {
+            "TRACKS": "TRK",
+            "CYLINDERS": "CYL"
+        };
+        return alcMap[getValue.toUpperCase()] || "TRK";
     }
 }