+++ packages/imperative/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## `8.13.0`
+- Format fix: `DeferredPromise` and `DeferredPromise.unit.test` comment format changed to match standard.
+
 ## `8.11.0`
 
 - Enhancement: Added `DeferredPromise` class to Imperative to provide utilities for data synchronization. [#2405](https://github.com/zowe/zowe-cli/pull/2405)